+++ gcpy/__init__.py
@@ -3,6 +3,7 @@ GCPy import script
 """
 
 from .benchmark import *
+from .community import *
 from .examples import *
 
 from .append_grid_corners import *
@@ -10,7 +11,6 @@ from .constants import *
 from .cstools import *
 from .date_time import *
 from .file_regrid import *
-from .format_hemco_data import *
 from .grid import *
 from .grid_stretching_transforms import *
 from .plot import *