+++ gcpy/community/__init__.py
@@ -1,0 +1,5 @@
+"""
+GCPy import script
+"""
+
+from .format_hemco_data import *